requirements.txt
streamlit==1.28.0
google-cloud-documentai==2.21.0
google-generativeai==0.3.2
python-dotenv==1.0.0
fastapi==0.103.2
uvicorn==0.23.2
requests==2.31.0
cachetools==5.3.2